Update to v0.0.4.
Remove chef_update in favor of a shell script as it seems to have issues when the build-essential cookbook is present.
Similarly, remove the Node.js cookbook in favor of a script that installs n and then node.js, as the binary installation with the cookbook doesn't work.
Update the protractor-selenium-server cookbook to v0.0.4 to deal with shifting Selenium jarfile URLs.
Update docs now that Protractor works with PhantomJS.